+++ Example/app/constants.js
