+++ Example/custom-templates/CustomTemplate.js
