+++ Example/index.js
@@ -3,7 +3,7 @@
  */
 
 import { AppRegistry } from 'react-native';
-import App from './App';
+import App from './app/App';
 import { name as appName } from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);
